[fix] type 정리 및 footer css fix
