+++ components/Payments/Order/FinalOrderSummary.module.scss
@@ -83,7 +83,7 @@ $divider: #f4f4f4;
     justify-content: space-between;
     align-items: center;
     padding: 14px 12px 4px;
-    margin-bottom: 88px;
+    padding-bottom: 88px;
 }
 
 .total_label {
