+++ constraint/order.ts
@@ -1,0 +1,7 @@
+const PRESETS = [
+    '요청사항 없음',
+    '문 앞에 놓아주세요',
+    '경비실에 맡겨 주세요',
+    '파손 위험 상품입니다. 배송 시 주의해주세요',
+    '직접 입력',
+] as const;
